requirements.txt
certifi==2022.9.24
minio==7.1.12
python-dotenv==0.21.0
urllib3==1.26.12